# openapi/openapi.yaml
...
openapi: 3.1.0
info:
  title: CafeOre POS API
  version: 1.0.0
  description: 珈琲・俺POSシステムのAPI仕様
servers:
  - url: http://localhost:8080
    description: 開発環境
paths:
...
  /api/items:
    get:
      summary: アイテム一覧取得
      operationId: getItems
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  $ref: '#/components/schemas/ItemResponse'
    post:
      summary: アイテム作成
      operationId: createItem
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ItemCreateRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemResponse'
  /api/items/{id}:
    get:
      summary: idからアイテム情報取得
      operationId: getItem
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemResponse'
    put:
      summary: アイテム情報更新
      operationId: updateItem
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ItemUpdateRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemResponse'
    delete:
      summary: アイテム削除
      operationId: deleteItem
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: 成功
  /api/menus:
    get:
      summary: メニュー一覧取得
      operationId: getMenus
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MenuResponse'
    post:
      summary: メニュー作成
      operationId: createMenu
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MenuCreateRequest'
      responses:
        '201':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MenuResponse'
  /api/menus/{id}:
    get:
      summary: メニュー取得
      operationId: getMenu
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MenuResponse'
    put:
      summary: メニュー更新
      operationId: updateMenu
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MenuUpdateRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MenuResponse'
    delete:
      summary: メニュー削除
      operationId: deleteMenu
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: 成功
  /api/item-types:
    get:
      summary: アイテムタイプ一覧取得
      operationId: getItemTypes
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  $ref: '#/components/schemas/ItemTypeResponse'
    post:
      summary: アイテムタイプ作成
      operationId: createItemType
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ItemTypeCreateRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemTypeResponse'
  /api/item-types/{id}:
    get:
      summary: idからアイテムタイプ情報取得
      operationId: getItemType
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemTypeResponse'
    put:
      summary: アイテム情報更新
      operationId: updateItemType
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ItemTypeUpdateRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemTypeResponse'
    delete:
      summary: アイテムタイプ削除
      operationId: deleteItemType
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: 成功
  /api/color-settings:
    get:
      summary: 背景色設定一覧取得
      operationId: getColorSettings
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ColorSettingResponse'
    put:
      summary: 背景色設定の作成・更新
      description: 対象（target_type, target_id）と画面（screen）の組が既にあれば色を上書きし、無ければ作成する
      operationId: upsertColorSetting
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ColorSettingUpsertRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ColorSettingResponse'
        '400':
          description: リクエストが不正
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/color-settings/{id}:
    delete:
      summary: 背景色設定削除
      operationId: deleteColorSetting
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: 成功
        '404':
          description: 背景色設定が見つかりません
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/orders:
    get:
      summary: オーダー一覧取得
      operationId: getOrders
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  $ref: '#/components/schemas/OrderResponse'
    post:
      summary: オーダー作成
      operationId: createOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderCreateRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
  /api/orders/{id}:
    get:
      summary: idからオーダー情報取得
      operationId: getOrder
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
    put:
      summary: オーダー情報更新
      operationId: updateOrder
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderUpdateRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
    delete:
      summary: オーダー削除
      operationId: deleteOrder
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: 成功
...
  /api/orders/{id}/served:
    patch:
      summary: オーダーを提供完了にする
      operationId: markOrderServe
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrderResponse'
        '404':
          description: オーダーが見つかりません
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/orders/{id}/cups/{cupId}/served:
    patch:
      summary: カップを提供完了にする
      description: 1杯ずつ提供済みと未提供を切り替える。全カップが提供済みになると注文も提供済みになり、外すと注文の提供済みも外れる。
      operationId: markOrderCupServe
      parameters:
        - name: id
          in: path
          required: true
          description: オーダーID
          schema:
            type: string
            format: uuid
        - name: cupId
          in: path
          required: true
          description: カップID（OrderResponse.cups[].id）
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '400':
          description: IDの形式が不正です
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: オーダーまたはカップが見つかりません
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/orders/{id}/comments:
    get:
      summary: 特定オーダーのコメント一覧取得
      operationId: getOrderComments
      tags:
        - comments
      parameters:
        - name: id
          in: path
          required: true
          description: オーダーID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CommentResponse'
        '404':
          description: オーダーが見つかりません
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: オーダーにコメント追加
      operationId: createOrderComment
      tags:
        - comments
      parameters:
        - name: id
          in: path
          required: true
          description: オーダーID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommentCreateRequest'
      responses:
        '201':
          description: 作成成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentResponse'
        '404':
          description: オーダーが見つかりません
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/caos/cups:
    put:
      summary: CaOS が決めたことを注文のカップに書く
      description: |
        CaOS（ドリップ管制）の盤面は、注文のカップ（OrderResponse の cups）の列（dripper・dripper_position・drip_id・brew_started_at・brew_finished_at）で持つ。
        割当・ドリッパーの移動・順番の入れ替え・未割当に戻す・統合は、どれもこの PUT でカップの列を書く。
        writes の全部を 1 つのトランザクションで書く（どれか 1 つでも通らなければ何も書かない）。
        書いたカップの注文は /api/ws/orders の {"type":"order"} で全部の画面に配り、ほかのインスタンスへは DB の通知（orders_changed）で知らせる。

        抽出の時刻は画面から送らない。空いているドリッパーに置いてそのまま始めるときは after の start_brew を true にし、サーバーが今の時刻を brew_started_at に入れる
        （iPad の時計がずれていても、残り時間がサーバーの時刻でそろう）。終えるのは「次へ」だけ。

        確かめること：
        - before が今の値と違う（ほかの端末が先に書いた・注文の編集で消えた・「次へ」で始まった）なら 409（楽観ロック）。画面は届いた注文で盤面を組み立て直す
        - 抽出中・終わりのカップ（brew_started_at のあるカップ）は書けない
        - 今日（日本時間）の注文のカップだけ書ける
        - 抽出が要らない種類（item_types.needs_brew が false）のカップは、ドリッパーにもカードにも入れられない
        - 指名の番号のあるカップ（明細の dripper が 1〜6）は、その番号のドリッパーにしか置けない
        - 限定のカップ（item_types.senior_only）は、今日の担当者が上級生のドリッパー（GET /api/caos/lanes の senior）にしか置けない。
          別のドリッパーへ置くとき（未割当から置く・ドリッパーを移す）だけ確かめ、同じドリッパーの中の順番の入れ替えは確かめない
          （担当者を上級生でない人に替えても、待っていた限定のカードはそのドリッパーに残るので）。
          指名のドリッパーの担当者が上級生でない限定のカップは、どこにも置けない
        - 1 つのドリッパーで同時に抽出中のカードは 1 枚。1 枚のカードは最大 2 杯。同じ drip_id のカップは同じ値（書かないカップも含めて）
        - 緊急のカップ（emergency_at のあるカップ）では、before・after の drip_id は入れ直しのカード（emergency_drip_id）を指す（最初に淹れた drip_id は書き換えない）。
          緊急のカップとほかのカップを 1 枚のカードにはできない
      operationId: writeCaosCups
      tags:
        - caos
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CaosCupsWriteRequest'
      responses:
        '204':
          description: 書いた
        '400':
          description: 形の違うリクエスト
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: before が今の値と違う（何も書かない）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: 決まりに合わない（何も書かない）。error を画面にそのまま出す
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/caos/drippers/{dripper}/next:
    post:
      summary: CaOS の「次へ」
      description: |
        そのドリッパーの抽出中のカードを終え（brew_finished_at）、そのカードのカップだけを準備完了にして（POS のカップの準備完了と同じく、注文の状態はカップから決め直す）、
        同じドリッパーの待機の先頭（dripper_position・注文番号の順）を抽出中にする（brew_started_at）。時刻はサーバーの今。
        抽出中が無ければ（マスターで準備完了にして終わった、など）待機の先頭を始めるだけ。
        drip_id には画面が抽出中と見ているカードを送る。今の抽出中と違えば 409（二度押しや、ほかの端末と同時に押したときに次のカードまで終わらせない）。
      operationId: advanceCaosDripper
      tags:
        - caos
      parameters:
        - name: dripper
          in: path
          required: true
          description: ドリッパーの番号（1〜6）
          schema:
            type: integer
            minimum: 1
            maximum: 6
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CaosNextRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CaosNextResult'
        '400':
          description: 形の違うリクエスト
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 画面の見ている抽出中が今と違う・抽出中も待機も無い（何も変えない）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/caos/emergency:
    post:
      summary: カップを緊急（入れ直し）にする
      description: |
        マスターの緊急ボタンと CaOS の入れ直しのパネルが使う。カップに緊急の印（emergency_at）を付けるだけで、入れ直しのカードは CaOS の未割当のいちばん上に出る
        （割当・「次へ」はふつうのカードと同じ。入れ直しのカードは emergency_drip_id で持ち、最初に淹れた drip_id は残す）。
        もう緊急のカップは何もしない（同じカップを 2 回緊急にしない）。カップの準備完了・提供済みは変えない（入れ直しのカードの「次へ」で準備完了にする）。

        interrupt が true なら、cup_ids のカップの抽出中のカードを中断する：そのカードのカップを全部緊急にし、カードは盤面から消え、
        そのドリッパーの待機の先頭を始める（時刻はサーバーの今）。interrupt が false でも、抽出中のカードのカップが全部緊急になれば同じく中断になる。

        緊急になったカップの注文には、プリンターにつないだレジが緊急のシール（「緊急」→ そのカップの本物と同じシール）を印刷する（emergency_printed_at）。
      operationId: markCaosEmergency
      tags:
        - caos
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CaosEmergencyRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CaosEmergencyResult'
        '400':
          description: 形の違うリクエスト
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: カップが消えた・注文の変更と重なった（何も変えない）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: 決まりに合わない（抽出の要らないカップ・今日でない注文・抽出中でないカードの中断。何も変えない）。error を画面にそのまま出す
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/orders/{id}/cups/{cupId}/emergency-label/claim:
    post:
      summary: 緊急のシールを印刷する役を取る
      description: |
        emergency_at があって emergency_printed_at が null のときだけ、emergency_printed_at にサーバーの今を付ける（条件付きの 1 回の書き込み）。
        付けられた（claimed が true）レジだけが緊急のシールを印刷する。レジが 2 台あっても、どちらか 1 台だけが付けられる。
        付けたら注文を配る（ほかの画面はもう印刷しない）。
      operationId: claimEmergencyLabel
      tags:
        - caos
      parameters:
        - name: id
          in: path
          required: true
          description: オーダーID
          schema:
            type: string
            format: uuid
        - name: cupId
          in: path
          required: true
          description: カップID（OrderResponse.cups[].id）
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 成功（付けられなかったときも 200 で claimed が false）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmergencyLabelClaim'
        '400':
          description: IDの形式が不正です
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: オーダーまたはカップが見つかりません
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/orders/{id}/cups/{cupId}/emergency-label/release:
    post:
      summary: 緊急のシールの印刷に失敗したので、印刷した時刻を空に戻す
      description: |
        emergency_printed_at が送った時刻（claim で付けた時刻）のままなら null に戻し、注文を配る（次の更新で、どれかのレジが試し直す）。
        ほかの値になっていれば何もしない（409）。
      operationId: releaseEmergencyLabel
      tags:
        - caos
      parameters:
        - name: id
          in: path
          required: true
          description: オーダーID
          schema:
            type: string
            format: uuid
        - name: cupId
          in: path
          required: true
          description: カップID（OrderResponse.cups[].id）
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmergencyLabelRelease'
      responses:
        '204':
          description: 空に戻した
        '400':
          description: 形の違うリクエスト
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: オーダーまたはカップが見つかりません
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 印刷した時刻が送った時刻と違う（何もしない）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/caos/lanes/{dripper}:
    put:
      summary: CaOS のドリッパーの担当者を替える（交代）
      description: |
        そのドリッパーの今日の担当者を替える。押したらその場で替える（抽出中かどうかは見ない。待機のカードもそのまま残す）。
        name が空なら担当者なし（senior は false にする）。senior は画面が sohosai-shift の上級生の名簿（seniors）で判定した値で、そのまま持つ。
        変えたら今日の担当者の全部を /api/ws/orders の {"type":"caos_lanes"} で全部の画面に配り、ほかのインスタンスへは DB の通知（caos_lanes_changed）で知らせる。
      operationId: putCaosLane
      tags:
        - caos
      parameters:
        - name: dripper
          in: path
          required: true
          description: ドリッパーの番号（1〜6）
          schema:
            type: integer
            minimum: 1
            maximum: 6
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CaosLaneUpdateRequest'
      responses:
        '200':
          description: 替えた。今日の担当者の全部
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CaosLanes'
        '400':
          description: 形の違うリクエスト
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/inventory/resources:
    post:
      summary: 在庫対象の作成
      operationId: createStockResource
      tags:
        - inventory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StockResourceRequest'
      responses:
        '201':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StockResourceResponse'
...
components:
  schemas:
...
    ItemResponse:
      type: object
      required:
        - id
        - name
        - abbr
        - item_type
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        abbr:
          type: string
        item_type:
          $ref: '#/components/schemas/ItemTypeResponse'
    # 作成リクエスト用（IDや自動生成フィールドを除外）
    ItemCreateRequest:
      type: object
      required:
        - name
        - abbr
        - item_type_id
      properties:
        name:
          type: string
        abbr:
          type: string
        item_type_id:
          type: string
          format: uuid
    # 更新リクエスト用
    ItemUpdateRequest:
      type: object
      required:
        - id
        - name
        - abbr
        - item_type_id
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        abbr:
          type: string
        item_type_id:
          type: string
          format: uuid
...
    MenuResponse:
      type: object
      required:
        - id
        - name
        - abbr
        - price
        - key
        - items
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        abbr:
          type: string
        price:
          type: integer
        key:
          type: string
        items:
          type: array
          items:
            $ref: '#/components/schemas/MenuItemResponse'
...
    MenuUpdateRequest:
      allOf:
        - $ref: '#/components/schemas/MenuCreateRequest'
        - type: object
          required:
            - id
          properties:
            id:
              type: string
              format: uuid
...